Refactor: split core project models and configs into dedicated files/packages
