--- core/src/main/kotlin/org/radarbase/core/model/Group.kt
@@ -1,10 +1,0 @@
-package org.radarbase.core.model
-
-import kotlinx.serialization.Serializable
-
-/** Group DTO returned by Management Portal groups API. */
-@Serializable
-data class Group(
-    val name: String,
-)
-
