+++ delegate-api/src/main/kotlin/org/radarbase/delegate/config/KeyStoreConfig.kt
@@ -1,0 +1,8 @@
+package org.radarbase.delegate.config
+
+data class KeyStoreConfig(
+    val alias: String,
+    val password: String,
+    val path: String,
+)
+
