+++ delegate-api/src/main/kotlin/org/radarbase/delegate/config/MPConfig.kt
@@ -1,0 +1,6 @@
+package org.radarbase.delegate.config
+
+data class MPConfig(
+    val url: String,
+)
+
