+++ delegate-api/src/main/kotlin/org/radarbase/delegate/config/PublicKeys.kt
@@ -1,0 +1,7 @@
+package org.radarbase.delegate.config
+
+data class PublicKeys(
+    val ecdsa: List<String>,
+    val rsa: List<String>,
+)
+
