+++ delegate-api/src/main/kotlin/org/radarbase/delegate/service/ProjectServiceImpl.kt
@@ -2,8 +2,8 @@ package org.radarbase.delegate.service
 
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
-import org.radarbase.core.model.Group
-import org.radarbase.core.model.Project
+import org.radarbase.core.model.project.Group
+import org.radarbase.core.model.project.Project
 import org.radarbase.delegate.config.DelegateConfig
 import org.radarbase.delegate.model.ProjectParticipant
 import org.slf4j.LoggerFactory
