+++ project-service/src/main/kotlin/org/radarbase/project/client/RadarProjectClient.kt
@@ -16,7 +16,7 @@ import jakarta.inject.Singleton
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
-import org.radarbase.core.model.Group
+import org.radarbase.core.model.project.Group
 import org.radarbase.core.util.KtorClientFactory
 import org.radarbase.core.util.ServiceTokenProvider
 import org.radarbase.project.config.ProjectServiceConfiguration
