+++ project-service/src/main/kotlin/org/radarbase/project/config/RadarConfiguration.kt
@@ -1,0 +1,15 @@
+package org.radarbase.project.config
+
+data class RadarConfiguration(
+    val baseUrl: String,
+    val timeout: Long,
+    val maxRetries: Int,
+    val auth: RadarAuthConfiguration,
+) {
+    data class RadarAuthConfiguration(
+        val enabled: Boolean,
+        val tokenHeader: String,
+        val tokenPrefix: String,
+    )
+}
+
