+++ project-service/src/main/kotlin/org/radarbase/project/service/ProjectService.kt
@@ -1,7 +1,7 @@
 package org.radarbase.project.service
 
-import org.radarbase.core.model.Group
-import org.radarbase.core.model.Project
+import org.radarbase.core.model.project.Group
+import org.radarbase.core.model.project.Project
 import org.radarbase.project.model.CreateGroupRequest
 import org.radarbase.project.model.ProjectParticipant
 import org.radarbase.project.model.RadarProject
