+++ project-service/src/main/kotlin/org/radarbase/project/service/ProjectServiceImpl.kt
@@ -2,18 +2,18 @@ package org.radarbase.project.service
 
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
-import org.radarbase.core.model.AgeRange
-import org.radarbase.core.model.Analysis
-import org.radarbase.core.model.Contact
-import org.radarbase.core.model.Design
-import org.radarbase.core.model.Domain
-import org.radarbase.core.model.Eligibility
-import org.radarbase.core.model.Group
-import org.radarbase.core.model.Intervention
-import org.radarbase.core.model.Outcomes
-import org.radarbase.core.model.Population
-import org.radarbase.core.model.Project
-import org.radarbase.core.model.Technology
+import org.radarbase.core.model.project.AgeRange
+import org.radarbase.core.model.project.Analysis
+import org.radarbase.core.model.project.Contact
+import org.radarbase.core.model.project.Design
+import org.radarbase.core.model.project.Domain
+import org.radarbase.core.model.project.Eligibility
+import org.radarbase.core.model.project.Group
+import org.radarbase.core.model.project.Intervention
+import org.radarbase.core.model.project.Outcomes
+import org.radarbase.core.model.project.Population
+import org.radarbase.core.model.project.Project
+import org.radarbase.core.model.project.Technology
 import org.radarbase.project.client.RadarProjectClient
 import org.radarbase.project.model.CreateGroupRequest
 import org.radarbase.project.model.ProjectParticipant
